refactor: Seat내 row, floor 타입 변경

+++ showhive-data/data/src/main/java/com/showhive/performance/domain/Seat.java
@@ -34,9 +34,9 @@ public class Seat extends BaseEntity {
 
     private String seatColumn;
 
-    private Integer seatRow;
+    private Short seatRow;
 
-    private Integer seatFloor;
+    private Short seatFloor;
 
     @NotBlank
     private String seatType;
